# src/schema/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ADA Soulbound Token API",
    "version": "1.0.0",
    "description": "Route definitions for ADA's Soulbound Token API"
  },
...
  "paths": {
...
    "/issuer/getAll": {
      "get": {
        "tags": [
          "Issuer"
        ],
        "summary": "Get all issuers",
        "description": "Get all issuers",
        "responses": {
          "200": {
            "description": "A list of issuers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Issuer"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/issuer/getAllByAddress": {
      "get": {
        "tags": [
          "Issuer"
        ],
        "summary": "Get all issuers by address",
        "description": "Get all issuers by address",
        "parameters": [
          {
            "in": "query",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string",
              "format": "address"
            },
            "description": "Issuer's address"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of issuers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Issuer"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/issuer/getIssuerByEmail": {
      "get": {
        "tags": [
          "Issuer"
        ],
        "summary": "Get issuer by email",
        "description": "Get issuer by email",
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Issuer email"
          }
        ],
        "responses": {
          "200": {
            "description": "An issuer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Issuer"
                }
              }
            }
          },
          "404": {
            "description": "Issuer not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/issuer/getIssuerByPK": {
      "get": {
        "tags": [
          "Issuer"
        ],
        "summary": "Get issuer by primary key",
        "description": "Get issuer by primary key",
        "parameters": [
          {
            "in": "query",
            "name": "issuerUuid",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Issuer primary key"
          }
        ],
        "responses": {
          "200": {
            "description": "An issuer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Issuer"
                }
              }
            }
          },
          "404": {
            "description": "Issuer not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/issuer/createNewIssuer": {
      "post": {
        "tags": [
          "Issuer"
        ],
        "summary": "Create new issuer",
        "description": "Create new issuer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "address": {
                    "type": "string",
                    "format": "address",
                    "example": "0xdD2FD4581271e330360230F9337D5c0430Bf44C0"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "mail@example.com"
                  },
                  "organization": {
                    "type": "string",
                    "example": "Some_Organization"
                  },
                  "name": {
                    "type": "string",
                    "example": "Name_of_the_issuer_for_the_event"
                  }
                },
                "required": [
                  "address",
                  "email",
                  "organization",
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Issuer created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Issuer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/issuer/updateIssuer": {
      "patch": {
        "tags": [
          "Issuer"
        ],
        "summary": "Update issuer",
        "description": "Update issuer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "issuerUuid": {
                    "type": "string",
                    "format": "uuid",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                  }
                },
                "required": [
                  "issuerUuid"
                ],
                "anyOf": [
                  {
                    "properties": {
                      "email": {
                        "type": "string",
                        "format": "email",
                        "example": "mail@example.com"
                      }
                    }
                  }
                ]
              }
            }
          },
          "responses": {
            "200": {
              "description": "Issuer updated successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Issuer"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid request"
            },
            "404": {
              "description": "Issuer not found"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      }
    },
...
    "/owner/getOwnerByAddress": {
      "get": {
        "tags": [
          "Owner"
        ],
        "summary": "Get owner by address",
        "description": "Get owner by address",
        "parameters": [
          {
            "in": "query",
            "name": "address",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Owner address"
          }
        ],
        "responses": {
          "200": {
            "description": "An owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Owner"
                }
              }
            }
          },
          "404": {
            "description": "Owner not found"
          }
        }
      }
    },
    "/owner/getOwnerByEmail": {
      "get": {
        "tags": [
          "Owner"
        ],
        "summary": "Get owner by email",
        "description": "Get owner by email",
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Owner email"
          }
        ],
        "responses": {
          "200": {
            "description": "An owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Owner"
                }
              }
            }
          },
          "404": {
            "description": "Owner not found"
          }
        }
      }
    },
    "/owner/getOwnerByPK": {
      "get": {
        "tags": [
          "Owner"
        ],
        "summary": "Get owner by primary key",
        "description": "Get owner by primary key",
        "parameters": [
          {
            "in": "query",
            "name": "ownerUuid",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Owner primary key"
          }
        ],
        "responses": {
          "200": {
            "description": "An owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Owner"
                }
              }
            }
          },
          "404": {
            "description": "Owner not found"
          }
        }
      }
    },
    "/owner/createOwner": {
      "post": {
        "tags": [
          "Owner"
        ],
        "summary": "Create new owner",
        "description": "Create new owner",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "address": {
                    "type": "string",
                    "format": "address",
                    "example": "0xdD2FD4581271e330360230F9337D5c0430Bf44C0"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "mail@example.com"
                  }
                },
                "required": [
                  "address",
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Owner created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Owner"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/owner/updateOwner": {
      "patch": {
        "tags": [
          "Owner"
        ],
        "summary": "Update owner",
        "description": "Update owner",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ownerUuid": {
                    "type": "string",
                    "format": "uuid",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                  }
                },
                "required": [
                  "ownerUuid"
                ],
                "anyOf": [
                  {
                    "properties": {
                      "email": {
                        "type": "string",
                        "format": "email",
                        "example": "mail@example.com"
                      }
                    }
                  },
                  {
                    "properties": {
                      "address": {
                        "type": "string",
                        "format": "address",
                        "example": "0xdD2FD4581271e330360230F9337D5c0430Bf44C0"
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Owner updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Owner"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Owner not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Issuer": {
        "type": "object",
        "properties": {
          "issuerUuid": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "issuerIdInContract": {
            "type": "integer",
            "example": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "mail@example.com"
          },
          "address": {
            "type": "string",
            "format": "address",
            "example": "0xdD2FD4581271e330360230F9337D5c0430Bf44C0"
          },
          "name": {
            "type": "string",
            "example": "Nerdearla"
          },
          "organization": {
            "type": "string",
            "example": "Programadores Argentinos"
          }
        }
      },
      "Owner": {
        "type": "object",
        "properties": {
          "ownerUuid": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "mail@example.com"
          },
          "address": {
            "type": "string",
            "format": "address",
            "example": "0xdD2FD4581271e330360230F9337D5c0430Bf44C0"
          }
        }
      },
...
    }
  }
}